+++ lib/service/ProdutoService.ts
@@ -17,11 +17,14 @@ export class ProdutoService {
 		});
 	}
 
-	public async update(id, produto: Produto): Promise<Produto> {
+	public async update(id: number, produto: Produto): Promise<Produto> {
 		return new Promise(async (resolve, reject) => {
-			const existe = await this.find({id});
+			if (!id) {
+				return reject(new BusinessException(getMessage(Messages.PARAMETROS_INVALIDOS)));
+			}
+			const existe = await this.find(id);
 			if (!existe) {
-				reject(new BusinessException(getMessage(Messages.REGISTRO_NAO_ENCONTRADO)));
+				return reject(new BusinessException(getMessage(Messages.REGISTRO_NAO_ENCONTRADO)));
 			}
 
 			await ProdutoModel.updateOne({id}, {
@@ -30,31 +33,34 @@ export class ProdutoService {
 				valor: produto.valor
 			}, (err, res) => {
 				if (err) {
-					reject(err);
+					return reject(err);
 				}
-				resolve(res);
+				return resolve(res);
 			});
 			reject(new BusinessException(getMessage(Messages.ERRO_INESPERADO)));
 		});
 	}
 
-	public async delete(id): Promise<void> {
+	public async delete(id: number): Promise<void> {
 		return new Promise(async (resolve, reject) => {
-			const existe = await this.find({id});
+			if (!id) {
+				return reject(new BusinessException(getMessage(Messages.PARAMETROS_INVALIDOS)));
+			}
+			const existe = await this.find(id);
 			if (!existe) {
-				reject(new BusinessException(getMessage(Messages.REGISTRO_NAO_ENCONTRADO)));
+				return reject(new BusinessException(getMessage(Messages.REGISTRO_NAO_ENCONTRADO)));
 			}
 
-			const permiteExcluir = await ServiceFactory.getItemVendaService().existeItemVendaProduto(id);
-			if (!permiteExcluir) {
-				reject(new BusinessException(getMessage(Messages.PRODUTO_POSSUI_VENDA)));
+			const possuiVenda = await ServiceFactory.getItemVendaService().existeItemVendaProduto(id);
+			if (possuiVenda) {
+				return reject(new BusinessException(getMessage(Messages.PRODUTO_POSSUI_VENDA)));
 			}
 
 			await ProdutoModel.deleteOne({id}, (err) => {
 				if (err) {
-					reject(err);
+					return reject(err);
 				}
-				resolve();
+				return resolve();
 			});
 			reject(new BusinessException(getMessage(Messages.ERRO_INESPERADO)));
 		});
@@ -64,7 +70,12 @@ export class ProdutoService {
 		return ProdutoModel.find();
 	}
 
-	public async find(id): Promise<Produto> {
-		return ProdutoModel.findOne({id});
+	public async find(id: number): Promise<Produto> {
+		return new Promise(async (resolve, reject) => {
+			if (!id) {
+				return reject(new BusinessException(getMessage(Messages.PARAMETROS_INVALIDOS)));
+			}
+			return resolve(await ProdutoModel.findOne({id}))
+		});
 	}
 }
